pyproject.toml
[project]
name = "dashboard"
version = "0.1.0"
description = "Personal productivity dashboard"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "google-api-python-client>=2.100",
    "google-auth>=2.23",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
